% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseasonalize.R
\name{deseasonalize}
\alias{deseasonalize}
\title{Deseasonalize a column in a dataframe}
\usage{
deseasonalize(data, .start = c(2003, 1), .frequency = 12)
}
\arguments{
\item{data}{A column with numerical data}

\item{.start}{Argument passed to seas function}

\item{.frequency}{Argument passed to seas function}
}
\description{
Deseasonalize a column in a dataframe
}
