% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hacienda.R
\name{theme_hacienda}
\alias{theme_hacienda}
\title{Theme similar to Hacienda's Excel theme}
\usage{
theme_hacienda(legend.position = "top")
}
\arguments{
\item{legend.position}{Position for the legend. The argument can be either a
string ("top", "bottom") or a vector with coordinates.}
}
\description{
Theme similar to Hacienda's Excel theme
}
