% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseasonalize.R
\name{deseasonalize}
\alias{deseasonalize}
\title{Deseasonalize a column in a dataframe}
\usage{
deseasonalize(data)
}
\arguments{
\item{data}{A column with numerical data}
}
\description{
Deseasonalize a column in a dataframe
}
